requirements.txt
Flask
Flask-Cors
PyPDF2
python-docx
